pyproject.toml
[tool.poetry]
name = "disney-route-optimize"
version = "0.1.0"
description = ""
authors = ["jumtra <data0science0acc@gmail.com>"]
readme = "README.md"
packages = [{include = "disney_route_optimize"}]

[tool.poetry.dependencies]
python = ">=3.10"
numpy = "^1.24.1"
pandas = "^1.5.3"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
ipykernel = "^6.21.1"
plotly = "^5.13.0"
nbformat = "^5.7.3"
selenium = "^4.10.0"
tqdm = "^4.65.0"
jpholiday = "^0.1.8"
omegaconf = "^2.3.0"
lightgbm = "^3.3.5"
mlflow = "^2.4.1"
torch = {version = "^2.0.1+cu117", source = "torch_cu117"}
torchvision = {version = "^0.15.2+cu117", source = "torch_cu117"}
torchaudio = {version = "^2.0.2+cu117", source = "torch_cu117"}
seaborn = "^0.12.2"
tslearn = "^0.6.1"
h5py = "^3.9.0"
pyclustering = "^0.10.1.2"
japanize-matplotlib = "^1.1.3"
pulp = "^2.7.0"


[[tool.poetry.source]]
name = "torch_cu117"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
include_trailing_comma = true
multi_line_output = 3
line_length = 150
profile = "black"
known_first_party = "disney_route_optimize/"

[tool.mypy]
files = [
    "disney_route_optimize",
]
ignore_missing_imports = true

[tool.black]
line-length=150